Release v0.2.25: preserve standings view types

+++ database/migrations/20260914_dashboard_results_and_standings.sql
@@ -26,36 +26,36 @@ with competition_scopes as (
   select
     home_team_registration_id as team_registration_id,
     competition_phase_id,
-    count(*)::integer as played,
-    count(*) filter (where home_score > away_score)::integer as won,
-    count(*) filter (where home_score = away_score)::integer as drawn,
-    count(*) filter (where home_score < away_score)::integer as lost,
-    coalesce(sum(home_score), 0)::integer as goals_for,
-    coalesce(sum(away_score), 0)::integer as goals_against
+    count(*) as played,
+    count(*) filter (where home_score > away_score) as won,
+    count(*) filter (where home_score = away_score) as drawn,
+    count(*) filter (where home_score < away_score) as lost,
+    coalesce(sum(home_score), 0) as goals_for,
+    coalesce(sum(away_score), 0) as goals_against
   from played_matches
   group by home_team_registration_id, competition_phase_id
   union all
   select
     away_team_registration_id as team_registration_id,
     competition_phase_id,
-    count(*)::integer as played,
-    count(*) filter (where away_score > home_score)::integer as won,
-    count(*) filter (where away_score = home_score)::integer as drawn,
-    count(*) filter (where away_score < home_score)::integer as lost,
-    coalesce(sum(away_score), 0)::integer as goals_for,
-    coalesce(sum(home_score), 0)::integer as goals_against
+    count(*) as played,
+    count(*) filter (where away_score > home_score) as won,
+    count(*) filter (where away_score = home_score) as drawn,
+    count(*) filter (where away_score < home_score) as lost,
+    coalesce(sum(away_score), 0) as goals_for,
+    coalesce(sum(home_score), 0) as goals_against
   from played_matches
   group by away_team_registration_id, competition_phase_id
 ), totals as (
   select
     team_registration_id,
     competition_phase_id,
-    sum(played)::integer as played,
-    sum(won)::integer as won,
-    sum(drawn)::integer as drawn,
-    sum(lost)::integer as lost,
-    sum(goals_for)::integer as goals_for,
-    sum(goals_against)::integer as goals_against
+    sum(played) as played,
+    sum(won) as won,
+    sum(drawn) as drawn,
+    sum(lost) as lost,
+    sum(goals_for) as goals_for,
+    sum(goals_against) as goals_against
   from team_results
   group by team_registration_id, competition_phase_id
 )
